[feat] FAQ 등록/수정 API 추가 및 식별자 Integer → Long 변경

+++ src/main/java/com/better/CommuteMate/attendance/controller/dto/AttendanceHistoryResponse.java
@@ -13,10 +13,10 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @NoArgsConstructor
 public class AttendanceHistoryResponse {
-    private Integer attendanceId;
+    private Long attendanceId;
     private LocalDateTime checkTime;
     private CodeType checkType;
-    private Integer scheduleId;
+    private Long scheduleId;
     private LocalDateTime scheduleStartTime;
     private LocalDateTime scheduleEndTime;
 }
