+++ src/main/java/com/better/CommuteMate/auth/application/dto/AuthTokens.java
@@ -5,14 +5,14 @@ import com.better.CommuteMate.global.code.CodeType;
 public class AuthTokens {
     private final String accessToken;
     private final String refreshToken;
-    private final long expiresAt;
-    private final Integer userId;
+    private final Long expiresAt;
+    private final Long userId;
     private final String userName;
     private final String email;
     private final CodeType roleCode;
 
-    public AuthTokens(String accessToken, String refreshToken, long expiresAt,
-                      Integer userId, String userName, String email, CodeType roleCode) {
+    public AuthTokens(String accessToken, String refreshToken, Long expiresAt,
+                      Long userId, String userName, String email, CodeType roleCode) {
         this.accessToken = accessToken;
         this.refreshToken = refreshToken;
         this.expiresAt = expiresAt;
@@ -30,11 +30,11 @@ public class AuthTokens {
         return refreshToken;
     }
 
-    public long getExpiresAt() {
+    public Long getExpiresAt() {
         return expiresAt;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
