+++ src/main/java/com/better/CommuteMate/auth/controller/dto/LoginResponse.java
@@ -14,7 +14,7 @@ public class LoginResponse extends ResponseDetail {
     private final String refreshToken;
     private final String tokenType;
     private final Long expiresAt;
-    private final Integer userId;
+    private final Long userId;
     private final String userName;
     private final String email;
     private final String roleCode;
